--- hooks/useDebounce.ts
@@ -1,18 +1,0 @@
-import { useEffect, useState } from 'react';
-
-const useDebounce = <T>(value: T, delay = 500): T => {
-  const [debouncedValue, setDebouncedValue] = useState(value);
-  useEffect(() => {
-    const handler = setTimeout(() => {
-      setDebouncedValue(value);
-    }, delay);
-
-    return () => {
-      clearTimeout(handler);
-    };
-  }, [value, delay]);
-
-  return debouncedValue;
-};
-
-export default useDebounce;
