+++ pages/search.tsx
@@ -1,5 +1,5 @@
 import { useDebounce } from '#hooks/index';
-import React, { useEffect, useState } from 'react';
+import { ChangeEvent, useEffect, useState } from 'react';
 import type { PostFields } from '#types/post';
 import PostCard from '#components/PostCard';
 
@@ -14,8 +14,6 @@ export default function Search() {
       const res = await fetch(`/api/search?keywords=${debouncedKeywords}`);
       const json = await res.json();
       setPosts(json.data);
-
-      console.log(json.data);
     };
 
     if (debouncedKeywords.replace(/\s/g, '').length) {
@@ -26,16 +24,13 @@ export default function Search() {
     setPosts([]);
   }, [debouncedKeywords]);
 
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>): void => {
+  const handleChange = (e: ChangeEvent<HTMLInputElement>): void => {
     const { value } = e.target;
     setKeywords(value);
   };
   return (
     <>
       <style jsx>{`
-        .thin-border-b {
-          border-bottom-width: 1px;
-        }
         .posts {
           grid-template-columns: repeat(auto-fill, minmax(12rem, 1fr));
         }
@@ -48,7 +43,7 @@ export default function Search() {
       `}</style>
       <div className="my-8">
         <input
-          className="thin-border-b w-full h-12 leading-tight border-0 border-gray-600 text-2xl focus:outline-none"
+          className="w-full h-12 leading-tight border-0 border-b border-gray-600 text-2xl focus:outline-none"
           type="text"
           placeholder="Dukun Teknologi Umum"
           onChange={handleChange}
